arriving time added
arriving time added for trains. so now we can directly get a current time then we will search for an available train, we already have an all the station names in the array.  it does not matter now from where the user is searching for a train cause we have added logic now "interval time" column